foodapp/requirements.txt
datasets
fastapi
gradio
loguru
numpy
Pillow
pre-commit
pydantic
python-multipart
pytorch-lightning
torchmetrics 
transformers
uvicorn[standard]
requests
gradio
tensorflow
